NHP DiffusionPreprocessing (#388)

+++ DiffusionPreprocessing/scripts/eddy_postproc.sh
@@ -14,25 +14,41 @@ SelectBestB0=$4 #0 only the actual diffusion data was fed into eddy
                 #1 least distorted b0 was prepended to the eddy input
                 # Note: This numeric value is used within the script as a numeric that controls
                 # the number of volumes to skip, so it isn't just used as 0/1 "boolean".
+BetSpeciesLabel=${5:-0} #0 = Human (uses standard 'bet'), >=1 = NHP species code passed to bet4animal:
+                        #  1=Chimp, 2=Macaque (rhesus/cyno/etc), 3=Marmoset, >3=other small NHPs.
+                        #  Controls bet fraction (0.3/0.4/0.5/0.6) and bet4animal -z label.
 
 globalscriptsdir=${HCPPIPEDIR_Global}
 
 eddydir=${workingdir}/eddy
 datadir=${workingdir}/data
 
-echo "Generating eddy QC report in ${workingdir}/QC"
-if [ -d "${workingdir}/QC" ]; then rm -r ${workingdir}/QC; fi
-qc_command=("${FSLDIR}/bin/eddy_quad")
-qc_command+=("${eddydir}/eddy_unwarped_images")
-qc_command+=(-idx "${eddydir}/index.txt")
-qc_command+=(-par "${eddydir}/acqparams.txt")
-qc_command+=(-m "${eddydir}/nodif_brain_mask.nii.gz")
-qc_command+=(-b "${eddydir}/Pos_Neg.bvals")
-qc_command+=(-g "${eddydir}/eddy_unwarped_images.eddy_rotated_bvecs")
-qc_command+=(-o "${workingdir}/QC")
-qc_command+=(-f "${workingdir}/topup/topup_Pos_Neg_b0_field.nii.gz")
-qc_command+=(-v)
-"${qc_command[@]}"
+# NHP-only: LSR resampling produces no JacobianResampling marker. In that case
+# eddy has already combined Pos/Neg, so skip eddy_quad QC and copy through.
+if [ "${BetSpeciesLabel}" -ne 0 ] && [ ! -e "${eddydir}/${EddyJacFlag}" ]; then
+	echo "LSR resampling has been used. Eddy Output has already been combined. CombineDataFlag is ignored."
+	cp ${eddydir}/Pos.bval ${datadir}/bvals
+	cp ${eddydir}/Pos.bvec ${datadir}/bvecs_noRot
+	cp ${eddydir}/eddy_unwarped_images.eddy_rotated_bvecs ${datadir}/bvecs_Rot_noLSR
+	cp ${eddydir}/eddy_unwarped_images.eddy_bvecs_for_SLR ${datadir}/bvecs
+	$FSLDIR/bin/imcp ${eddydir}/eddy_unwarped_images ${datadir}/data
+	mkdir -p ${workingdir}/QC
+	echo "Eddy QC report is not generated for LSR resampling"
+else
+	echo "Generating eddy QC report in ${workingdir}/QC"
+	if [ -d "${workingdir}/QC" ]; then rm -r ${workingdir}/QC; fi
+	qc_command=("${FSLDIR}/bin/eddy_quad")
+	qc_command+=("${eddydir}/eddy_unwarped_images")
+	qc_command+=(-idx "${eddydir}/index.txt")
+	qc_command+=(-par "${eddydir}/acqparams.txt")
+	qc_command+=(-m "${eddydir}/nodif_brain_mask.nii.gz")
+	qc_command+=(-b "${eddydir}/Pos_Neg.bvals")
+	qc_command+=(-g "${eddydir}/eddy_unwarped_images.eddy_rotated_bvecs")
+	qc_command+=(-o "${workingdir}/QC")
+	qc_command+=(-f "${workingdir}/topup/topup_Pos_Neg_b0_field.nii.gz")
+	qc_command+=(-v)
+	"${qc_command[@]}"
+fi
 
 #Prepare for next eddy Release
 #if [ ! -e ${eddydir}/${EddyJacFlag} ]; then
@@ -151,6 +167,11 @@ if [ ! $GdCoeffs = "NONE" ]; then
 	# Transform CNR maps
 	${CARET7DIR}/wb_command -volume-dilate ${warpedDir}/cnr_maps_warped.nii.gz $DilateDistance NEAREST ${warpedDir}/cnr_maps_dilated.nii.gz
 	${FSLDIR}/bin/applywarp --rel --interp=spline -i ${warpedDir}/cnr_maps_dilated -r ${warpedDir}/cnr_maps_dilated -w ${warpedDir}/fullWarp -o ${datadir}/cnr_maps
+	# NHP-only: spline interpolation can produce NaN at borders for low-SNR data;
+	# fall back to sinc interpolation if any NaN is present in cnr_maps
+	if [ "${BetSpeciesLabel}" -ne 0 ] && [[ $(${FSLDIR}/bin/fslstats ${datadir}/cnr_maps.nii.gz -m) =~ nan ]]; then
+		${FSLDIR}/bin/applywarp --rel --interp=sinc -i ${warpedDir}/cnr_maps_dilated -r ${warpedDir}/cnr_maps_dilated -w ${warpedDir}/fullWarp -o ${datadir}/cnr_maps
+	fi
 	${FSLDIR}/bin/imrm ${warpedDir}/cnr_maps_dilated
 
 	# Transform field of view mask (using conservative trilinear interpolation with high threshold)
@@ -176,6 +197,18 @@ ${FSLDIR}/bin/fslmaths ${datadir}/cnr_maps -mas ${datadir}/fov_mask ${datadir}/c
 # Remove negative intensity values (from eddy) from final data
 ${FSLDIR}/bin/fslmaths ${datadir}/data -thr 0 ${datadir}/data
 ${FSLDIR}/bin/fslroi ${datadir}/data ${datadir}/nodif 0 1
-${FSLDIR}/bin/bet ${datadir}/nodif ${datadir}/nodif_brain -m -f 0.1
+if [ "${BetSpeciesLabel}" -eq 0 ]; then
+	# Human: standard FSL bet
+	${FSLDIR}/bin/bet ${datadir}/nodif ${datadir}/nodif_brain -m -f 0.1
+else
+	# NHP: bet4animal with species-tuned fraction
+	case "${BetSpeciesLabel}" in
+		1) BetFraction=0.3 ;;
+		2) BetFraction=0.4 ;;
+		3) BetFraction=0.5 ;;
+		*) BetFraction=0.6 ;;
+	esac
+	${FSLDIR}/bin/bet4animal ${datadir}/nodif ${datadir}/nodif_brain -m -f ${BetFraction} -z ${BetSpeciesLabel}
+fi
 
 echo -e "\n END: eddy_postproc"
